Estimator now scores after checking particle collision number.

+++ Transport.cpp
@@ -15,84 +15,74 @@ void Transport::runTransport()
 {
     numHis = constants->getNumHis();
     double tally = 0;
-    for(int i = 0; i < numHis ; i++)
+    for( int i = 0; i < numHis; i++ )
     {
         //start a timer
         timer->startHist();
 
         //sample src 
         Part_ptr p_new = geometry->sampleSource();
-        //Part_ptr p_new = make_shared<Particle>(point(0,0,0), point(0,0,1), 1);
         Cell_ptr startingCell = geometry->whereAmI(p_new->getPos());
         p_new->setCell(startingCell);
         pstack.push(p_new);
           
         //run history
-        while(!pstack.empty())
+        while( !pstack.empty() )
         {
            Part_ptr p = pstack.top();
             while(p->isAlive())
             {
-            //p->printState();
                 Cell_ptr current_Cell = p->getCell();
 
                 double d2s = current_Cell->distToSurface(p);
                 double d2c = current_Cell->distToCollision(p);
-            //cout << "d2s: " << d2s << "  d2c: " << d2c << endl;
                 
                 if(d2s > d2c) //collision!
                 {
+                    // move particle
+                    p->move(d2c);
+
                     // score collision tally in current cell
                     timer->startTimer("scoring collision tally");
-                    current_Cell->scoreTally(p , current_Cell->getMat()->getMacroXS( p ) ); 
-                    tally++;
+                    current_Cell->scoreTally( p, current_Cell->getMat()->getMacroXS( p ) );
                     timer->endTimer("scoring collision tally");
 
-                    timer->startTimer("scoring mesh tally");
-                    //std::cout << "About to score mesh tally " << std::endl;
                     // score mesh tally
-                    mesh->scoreTally( p , current_Cell->getMat()->getMacroXS( p ) );
-                    //std::cout << "We scored that mesh tally! " << std::endl;
+                    timer->startTimer("scoring mesh tally");
+                    mesh->scoreTally( p, current_Cell->getMat()->getMacroXS( p ) );
                     timer->endTimer("scoring mesh tally");
+                    tally++;
 
-                    p->move(d2c);
+                    // sample collision
                     current_Cell->getMat()->sampleCollision( p, pstack );
-                    p->kill(); //TODO: make this not awful
+
+                    // check number of collisions
+                    if ( p->getNCollisions() >= constants->getKillAfterNColl() ) 
+                    {
+                        p->kill(); 
+                    }
                 }
                 else //hit surface
                 {
+                    p->move(d2s + 0.00000001); // TODO: convert this to tolerance with Constants
+                    Cell_ptr newCell = geometry->whereAmI( p->getPos() );
 
-                    p->move(d2s + 0.00000001);
-                    Cell_ptr newCell = geometry->whereAmI(p->getPos());
-                if(newCell == nullptr)
-                {
-                    p->kill();
-                }
-                else
-                {
-                    p->setCell(newCell);
-                }
+                    if ( newCell == nullptr ) { p->kill(); }
+                    else { p->setCell(newCell); }
                 }
             }
             pstack.pop();
         }
         //tell all estimators that the history has ended
-         for( auto cell : geometry->getCells() ) {
-        cell->endTallyHist();
-         }
+        for( auto cell : geometry->getCells() ) { cell->endTallyHist(); }
 
-           // end histories in the mesh
-           mesh->endTallyHist();
+        // end histories in the mesh
+        mesh->endTallyHist();
 
         // end the history timer
         timer->endHist();
     }
-
-    cout << std::endl << "Transport finished!" << std::endl;
-    cout << std::endl << "************************************************************************" << std::endl;
-    cout << "************************************************************************" << std::endl;
     tally /= numHis;
-    //cout << "tally " << tally << endl;
 }
 
 void Transport::output() {
